редирект при прямом входе не залогиненого пользователя
